lib: path: move path.hpp in include/path/
