+++ applications/cup2021/path.cpp
@@ -2,7 +2,7 @@
 #include <kernel_defines.h>
 
 #include "path/Pose.hpp"
-#include "path.hpp"
+#include "path/path.hpp"
 #include "platform.hpp"
 
 // TODO:
