+++ lib/path/path.cpp
@@ -1,4 +1,4 @@
-#include "path.hpp"
+#include "path/path.hpp"
 
 #include "platform.hpp"
 #include "utils.h"
